database_wcp/pyproject.toml
[project]
name = "trash-optimizer-data-scripts"
version = "1.0.0"
description = "Data extraction and processing scripts for Trash Optimizer"
requires-python = ">=3.10"
dependencies = [
    "google-cloud-bigquery>=3.11.0",
    "google-auth>=2.23.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
extract-garbage-locations = "extract_garbage_location_nantes:main"
create-trash-dataset = "create_trash_dataset:main"
create-trash-points = "create_trash_points:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []
